requirements_dev.txt
coverage>=4.5.3
psutil>=5.6.3
PyYAML>=5.1.1
requests>=2.22.0
bottle>=0.12.17
mailtest>=1.1.3
